requirements.txt
fpdf==1.7.2
openai==1.53.0
PyPDF2==3.0.1
python-dotenv==1.0.1
